refactor: 채널 목록 조회 로직 ChannelSubscriptionService에서 ChannelService로 이동
