+++ backend/src/main/java/com/pickpick/channel/application/ChannelSubscriptionService.java
@@ -5,8 +5,6 @@ import com.pickpick.channel.domain.ChannelRepository;
 import com.pickpick.channel.domain.ChannelSubscription;
 import com.pickpick.channel.domain.ChannelSubscriptionRepository;
 import com.pickpick.channel.ui.dto.ChannelOrderRequest;
-import com.pickpick.channel.ui.dto.ChannelResponse;
-import com.pickpick.channel.ui.dto.ChannelResponses;
 import com.pickpick.channel.ui.dto.ChannelSubscriptionRequest;
 import com.pickpick.exception.channel.ChannelNotFoundException;
 import com.pickpick.exception.channel.SubscriptionDuplicateException;
@@ -38,27 +36,6 @@ public class ChannelSubscriptionService {
         this.channelSubscriptions = channelSubscriptions;
         this.channels = channels;
         this.members = members;
-    }
-
-    public ChannelResponses findAll(final Long memberId) {
-        List<Channel> allChannels = channels.findAllByOrderByName();
-        Map<Long, Channel> subscribedChannels = findSubscribedChannels(memberId);
-
-        List<ChannelResponse> channelResponses = findChannelResponses(allChannels, subscribedChannels);
-        return new ChannelResponses(channelResponses);
-    }
-
-    private List<ChannelResponse> findChannelResponses(final List<Channel> allChannels,
-                                                       final Map<Long, Channel> subscribedChannels) {
-        return allChannels.stream()
-                .map(channel -> ChannelResponse.of(subscribedChannels, channel))
-                .collect(Collectors.toList());
-    }
-
-    private Map<Long, Channel> findSubscribedChannels(final Long memberId) {
-        return channelSubscriptions.findAllByMemberId(memberId)
-                .stream()
-                .collect(Collectors.toMap(ChannelSubscription::getChannelId, ChannelSubscription::getChannel));
     }
 
     public List<ChannelSubscription> findAllOrderByViewOrder(final Long memberId) {
