+++ backend/src/main/java/com/pickpick/channel/ui/ChannelController.java
@@ -1,7 +1,7 @@
 package com.pickpick.channel.ui;
 
 import com.pickpick.auth.support.AuthenticationPrincipal;
-import com.pickpick.channel.application.ChannelSubscriptionService;
+import com.pickpick.channel.application.ChannelService;
 import com.pickpick.channel.ui.dto.ChannelResponses;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,14 +11,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/channels")
 public class ChannelController {
 
-    private final ChannelSubscriptionService channelSubscriptionService;
+    private final ChannelService channelService;
 
-    public ChannelController(final ChannelSubscriptionService channelSubscriptionService) {
-        this.channelSubscriptionService = channelSubscriptionService;
+    public ChannelController(final ChannelService channelService) {
+        this.channelService = channelService;
     }
 
     @GetMapping
     public ChannelResponses findAllChannels(final @AuthenticationPrincipal Long memberId) {
-        return channelSubscriptionService.findAll(memberId);
+        return channelService.findAll(memberId);
     }
 }
